moved Database class to util package
